Add a stylesheet file, a "Get Started" string and change the index page
So I've added basic styling. This is by far not done, but already looks
a lot better. For lack of decision, I used Source Sans Pro (which I kinda like
as a font). It's all black on white for the moment.

The footer still has a bug, as it's always on the bottom of the viewport. Ah
well, CSS.